Making the source build idempotent

+++ tools/ubuntu-provision-source.sh
@@ -9,15 +9,25 @@ sudo apt-get update -qq
 sudo apt-get build-dep -yqq wireshark
 sudo apt-get install -yqq cmake qt5-default libqt5multimedia5 qtmultimedia5-dev qttools5-dev qttools5-dev-tools
 
-# Check out source and patch
+# Check out source
 mkdir -p ${HOME}/wireshark-$WIRESHARK_BRANCH
 cd ${HOME}/wireshark-$WIRESHARK_BRANCH
 git init
-git remote add -t $WIRESHARK_BRANCH -f origin https://github.com/wireshark/wireshark;
+if ! git config remote.origin.url > /dev/null; then
+  git remote add -t $WIRESHARK_BRANCH -f origin https://github.com/wireshark/wireshark;
+fi
 git checkout $WIRESHARK_BRANCH
-ln -s $PLUGIN_DIR plugins/sap
-git apply $PLUGIN_DIR/wireshark-$WIRESHARK_BRANCH.patch
+
+# Link the plugin to the plugins directory if required
+if [ ! -e plugins/sap ]; then
+  ln -s $PLUGIN_DIR plugins/sap
+fi
+
+# Apply the patch if required
+if git apply --check $PLUGIN_DIR/wireshark-$WIRESHARK_BRANCH.patch; then
+  git apply $PLUGIN_DIR/wireshark-$WIRESHARK_BRANCH.patch
+fi
 
 # Install test requirements
-sudo apt-get install -yqq libxml2-dev libxslt-dev python-dev
+sudo apt-get install -yqq libxml2-dev libxslt-dev python-dev python-pip
 sudo -H sh -c "CXX=g++ CC=gcc pip install pysap pyshark"
